Move stats to feature folder
